+++ forecasting_tools/auto_optimizers/bot_evaluator.py
@@ -13,12 +13,12 @@ from forecasting_tools.auto_optimizers.prompt_data_models import (
     EvaluatedBot,
     PromptIdea,
 )
-from forecasting_tools.benchmarking.benchmark_for_bot import BenchmarkForBot
-from forecasting_tools.benchmarking.benchmarker import Benchmarker
-from forecasting_tools.data_models.question_plus_research import (
+from forecasting_tools.auto_optimizers.question_plus_research import (
     QuestionPlusResearch,
     ResearchType,
 )
+from forecasting_tools.benchmarking.benchmark_for_bot import BenchmarkForBot
+from forecasting_tools.benchmarking.benchmarker import Benchmarker
 from forecasting_tools.data_models.questions import MetaculusQuestion
 from forecasting_tools.forecast_bots.forecast_bot import ForecastBot
