+++ scripts/run_benchmarker.py
@@ -6,10 +6,10 @@ import logging
 from forecasting_tools.ai_models.resource_managers.monetary_cost_manager import (
     MonetaryCostManager,
 )
-from forecasting_tools.benchmarking.benchmarker import Benchmarker
-from forecasting_tools.data_models.question_plus_research import (
+from forecasting_tools.auto_optimizers.question_plus_research import (
     QuestionPlusResearch,
 )
+from forecasting_tools.benchmarking.benchmarker import Benchmarker
 from forecasting_tools.forecast_bots.forecast_bot import ForecastBot
 from forecasting_tools.util.custom_logger import CustomLogger
 from run_bots import configure_and_run_bot, get_all_bots
